refactor(#451): controller dir
